Added nasty metachars to config file
New function translate_escapes() to handle escape sequences in
metachars.  I left the debug logging in for potential interested
testers.  This patch most likely cannot handle UTF8, but I doubt nrpe
handled it before hand either.

The decision to make the nasty_metachars configurable was made to try to make
it easier for users to add the newline, or remove it.  I don't know if
newline should be added to the default list of nasty metas or not.  At
least not without a bit of warning, as some checks in the wild may
require it.

Only a few escaped control chars are supported at this point.  Let me
know if:

1. there are others that should be added
2. there is a better way to do it
3. there is a way to include UTF8 support?